+++ CHANGELOG.md
@@ -5,6 +5,8 @@ current version: 0.1.1
 ### 0.1.1 | 2020-11-30
 
 - [变更] 组件使用 implitcitWidth/Height
+- [新增] PyHandler.register() 装饰器
+- [更新] LCCheckList 和 LCRadioList 修改代理组件属性的新方案
 
 ### 0.1.0 | 2020-11-28